Rename rothko to helpers
